デザイン途中

+++ app/Http/Controllers/StylistController.php
@@ -21,110 +21,7 @@ class StylistController extends Controller
         return view('stylist/homeStylist');
     }
     
-    /**
-     * ------------------------美容師情報関連------------------------
-     */
-    
-    
-   /**
-     * 美容師情報を新規登録
-     * 
-     */
-    //  public function create(User $user)
-    //  {
-    //     $user = Auth::user();
-    //     return view('stylist/info/create')->with(['user' => $user]);
-        
-    //  }
-     
-    /**
-     * 新規作成した投稿をDBに反映
-     * @param Request $request
-     * @return Response
-     */
-    //     public function store(User $user, UserRequest $request)
-    // {
-    //     //formのnameが'user[**]'の入力を$articleに格納
-    //     $input = $request['user'];
-    //     //formにpasswordを挿入
-    //     $input['password'] = Hash::make($request['password']);
-        
-    //     //画像ファイルは$request-+>fileで受け取る
-    //     $photo = $request->file('image');
-    //     $photo_name = $photo->getClientOriginalName();
-    //     //保存するディレクトリ(storage/app/public以下)、ファイル、ファイル名の順
-    //     Storage::disk('public')->putFileAs('stylist',$photo,$photo_name);
-    //     $input['image'] = $photo_name;
-    //     $user->fill($input)->save();
-        
-    //     return redirect('/account/' . $user->id);
-    // }
-    
-    /**
-     * 美容師情報を表示
-     */
-    
-    //   public function showAccount(User $user, Menu $menu, Catalog $catalog)
-    // {
-    //     //メニュー情報取得
-    //     $menus=$user->menus()->get();
-    //     //カタログ情報取得
-    //     $catalogs=$user->catalogs()->get();
-    //     //ユーザー情報取得
-    //     $user = Auth::user();
-        
-    //     return view('stylist/showAccount')->with([
-    //         'user' => $user,
-    //         'menus' => $menus,
-    //         'catalogs' => $catalogs,
-    //         ]);
-    // }
-    
-    /**
-     * 美容師情報を編集
-     * 
-     */
-        public function edit(User $user)
-    {
-        $user = Auth::user();
-        return view('stylist/info/edit')->with(['user' => $user]);
-    }
-    
-    /**
-     * 美容師情報を更新
-     * 
-     */
-        public function update(UserRequest $request, User $user)
-    {
-       //formのnameが'user[**]'の入力を$inputに格納
-        $input_user = $request['user'];
-        
-        //画像ファイルを変更するとき
-        if($request->hasFile('image')) {
-            Storage::delete('public/stylist/' . $user->image); //元の画像を削除☆
-            $photo = $request->file('image');
-            $photo_name = $photo->getClientOriginalName();
-            //保存するディレクトリ(storage/app/public以下)、ファイル、ファイル名の順
-            Storage::disk('public')->putFileAs('stylist',$photo,$photo_name);
-            $input_user['image'] = $photo_name;
-            $user->fill($input_user)->save();
-        }else{
-             //画像ファイルを変更しない時
-            $user->fill($input_user)->save(); //image以外を保存☆☆
-        }
-        
-        return redirect('/account/' . $user->id);
-    }
-    
-    /**
-     * 美容師情報を削除
-     * 
-     */
-        public function delete(User $user)
-    {
-        $user->delete();
-        return redirect('/account');
-    }
+ 
     
     /**
      * ------------------------メニュー情報関連(menus)------------------------
@@ -169,6 +66,9 @@ class StylistController extends Controller
      */
     public function editMenu(User $user, Menu $menu)
     {
+        // update, destroyでも同様に
+        $this->authorize('edit', $menu);
+        
         return view('stylist/menu/EditMenu')->with([
                 'menu' => $menu,
                 ]);
@@ -179,6 +79,9 @@ class StylistController extends Controller
      */
     public function updateMenu(Request $request, User $user, Menu $menu)
     {
+        // update, destroyでも同様に
+        $this->authorize('edit', $menu);
+        
         $user = Auth::user();
         $input_menu_update = $request['menu'];
         $input_menu_update['user_id']=$user->id;
@@ -192,6 +95,9 @@ class StylistController extends Controller
      */
         public function deleteMenu(User $user, Menu $menu)
     {
+        // update, destroyでも同様に
+        $this->authorize('edit', $menu);
+        
         $user = Auth::user();
         $menu->delete();
         return redirect('/account/');
